[tooltip][previewcard][popover] Refactor viewport helper functions into a shared utility module
