+++ packages/react/src/preview-card/viewport/PreviewCardViewport.tsx
@@ -5,6 +5,11 @@ import { useAnimationFrame } from '@base-ui/utils/useAnimationFrame';
 import { usePreviousValue } from '@base-ui/utils/usePreviousValue';
 import { useIsoLayoutEffect } from '@base-ui/utils/useIsoLayoutEffect';
 import { useStableCallback } from '@base-ui/utils/useStableCallback';
+import {
+  calculateRelativePosition,
+  getActivationDirection,
+  Offset,
+} from '@base-ui/utils/viewportHelpers';
 import { usePreviewCardRootContext } from '../root/PreviewCardContext';
 import { usePreviewCardPositionerContext } from '../positioner/PreviewCardPositionerContext';
 import { BaseUIComponentProps } from '../../utils/types';
@@ -250,74 +255,6 @@ export namespace PreviewCardViewport {
     transitioning: boolean;
     instant: 'dismiss' | 'focus' | undefined;
   }
-}
-
-type Offset = {
-  horizontal: number;
-  vertical: number;
-};
-
-/**
- * Returns a string describing the provided offset.
- * It describes both the horizontal and vertical offset, separated by a space.
- *
- * @param offset
- */
-function getActivationDirection(offset: Offset | null): string | undefined {
-  if (!offset) {
-    return undefined;
-  }
-
-  return `${getValueWithTolerance(offset.horizontal, 5, 'right', 'left')} ${getValueWithTolerance(offset.vertical, 5, 'down', 'up')}`;
-}
-
-/**
- * Returns a label describing the value (positive/negative) treating values
- * within tolerance as zero.
- *
- * @param value Value to check
- * @param tolerance Tolerance to treat the value as zero.
- * @param positiveLabel
- * @param negativeLabel
- * @returns If 0 < abs(value) < tolerance, returns an empty string. Otherwise returns positiveLabel or negativeLabel.
- */
-function getValueWithTolerance(
-  value: number,
-  tolerance: number,
-  positiveLabel: string,
-  negativeLabel: string,
-) {
-  if (value > tolerance) {
-    return positiveLabel;
-  }
-
-  if (value < -tolerance) {
-    return negativeLabel;
-  }
-
-  return '';
-}
-
-/**
- * Calculates the relative position between centers of two elements.
- */
-function calculateRelativePosition(from: Element, to: Element): Offset {
-  const fromRect = from.getBoundingClientRect();
-  const toRect = to.getBoundingClientRect();
-
-  const fromCenter = {
-    x: fromRect.left + fromRect.width / 2,
-    y: fromRect.top + fromRect.height / 2,
-  };
-  const toCenter = {
-    x: toRect.left + toRect.width / 2,
-    y: toRect.top + toRect.height / 2,
-  };
-
-  return {
-    horizontal: toCenter.x - fromCenter.x,
-    vertical: toCenter.y - fromCenter.y,
-  };
 }
 
 /**
