+++ packages/utils/src/viewportHelpers.ts
@@ -1,0 +1,67 @@
+export type Offset = {
+  horizontal: number;
+  vertical: number;
+};
+
+/**
+ * Returns a string describing the provided offset.
+ * It describes both the horizontal and vertical offset, separated by a space.
+ *
+ * @param offset
+ */
+export function getActivationDirection(offset: Offset | null): string | undefined {
+  if (!offset) {
+    return undefined;
+  }
+
+  return `${getValueWithTolerance(offset.horizontal, 5, 'right', 'left')} ${getValueWithTolerance(offset.vertical, 5, 'down', 'up')}`;
+}
+
+/**
+ * Returns a label describing the value (positive/negative) trating values
+ * within tolarance as zero.
+ *
+ * @param value Value to check
+ * @param tolerance Tolerance to treat the value as zero.
+ * @param positiveLabel
+ * @param negativeLabel
+ * @returns If 0 < abs(value) < tolerance, returns an empty string. Otherwise returns positiveLabel or negativeLabel.
+ */
+export function getValueWithTolerance(
+  value: number,
+  tolerance: number,
+  positiveLabel: string,
+  negativeLabel: string,
+) {
+  if (value > tolerance) {
+    return positiveLabel;
+  }
+
+  if (value < -tolerance) {
+    return negativeLabel;
+  }
+
+  return '';
+}
+
+/**
+ * Calculates the relative position between centers of two elements.
+ */
+export function calculateRelativePosition(from: Element, to: Element): Offset {
+  const fromRect = from.getBoundingClientRect();
+  const toRect = to.getBoundingClientRect();
+
+  const fromCenter = {
+    x: fromRect.left + fromRect.width / 2,
+    y: fromRect.top + fromRect.height / 2,
+  };
+  const toCenter = {
+    x: toRect.left + toRect.width / 2,
+    y: toRect.top + toRect.height / 2,
+  };
+
+  return {
+    horizontal: toCenter.x - fromCenter.x,
+    vertical: toCenter.y - fromCenter.y,
+  };
+}
